backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
python-dotenv
openai
requests
numpy
gunicorn==21.2.0